Add click listener for the steps.

+++ app/src/main/java/io/github/yahia_hassan/bakingapp/POJO/Ingredient.java
@@ -4,18 +4,18 @@ import android.os.Parcel;
 import android.os.Parcelable;
 
 public class Ingredient implements Parcelable {
-    private int mQuantity;
+    private double mQuantity;
     private String mMeasure;
     private String mIngredient;
 
-    public Ingredient(int quantity, String measure, String ingredient) {
+    public Ingredient(double quantity, String measure, String ingredient) {
         mQuantity = quantity;
         mMeasure = measure;
         mIngredient = ingredient;
     }
 
     protected Ingredient(Parcel in) {
-        mQuantity = in.readInt();
+        mQuantity = in.readDouble();
         mMeasure = in.readString();
         mIngredient = in.readString();
     }
@@ -39,12 +39,12 @@ public class Ingredient implements Parcelable {
 
     @Override
     public void writeToParcel(Parcel dest, int flags) {
-        dest.writeInt(mQuantity);
+        dest.writeDouble(mQuantity);
         dest.writeString(mMeasure);
         dest.writeString(mIngredient);
     }
 
-    public int getQuantity() {
+    public double getQuantity() {
         return mQuantity;
     }
